style(client): rename files
