+++ app/src/components/Slate/Richtext.tsx
@@ -10,7 +10,7 @@ import {
 } from "slate";
 import { withHistory } from "slate-history";
 
-import { Button, Icon, Toolbar } from "./components";
+import { Button, Icon, Toolbar } from "./Components";
 
 const HOTKEYS: { [x: string]: string } = {
   "mod+b": "bold",
